fourth

+++ db/migrate/20220510060639_remove_p_id_from_task.rb
@@ -1,5 +1,5 @@
 class RemovePIdFromTask < ActiveRecord::Migration[7.0]
   def change
-    remove_column :tasks, :p_id, :decimal
+    remove_column :tasks, :p_id, :integer
   end
 end
